requirements.txt
Django
djangorestframework
django-cors-headers
Pillow
django-ace
whitenoise
gunicorn
psycopg2-binary
dj-database-url
